Move SetAssignmentsModal to separate file
