+++ src/app/components/pages/SetAssignments.tsx
@@ -1,4 +1,4 @@
-import React, {ChangeEvent, useCallback, useEffect, useMemo, useState} from "react";
+import React, {useCallback, useEffect, useMemo, useState} from "react";
 import {
     Alert,
     Button,
@@ -6,9 +6,6 @@ import {
     CardBody,
     Col,
     Container,
-    Form,
-    FormFeedback,
-    FormGroup,
     Input,
     Label,
     Row,
@@ -18,7 +15,6 @@ import {
 } from "reactstrap";
 import {Link, useLocation} from "react-router-dom";
 import {
-    assignGameboard,
     closeActiveModal,
     openActiveModal,
     openIsaacBooksModal,
@@ -31,10 +27,7 @@ import {
     useUnassignGameboardMutation
 } from "../../state";
 import {ShowLoading} from "../handlers/ShowLoading";
-import range from "lodash/range";
-import sortBy from "lodash/sortBy";
 import {TitleAndBreadcrumb} from "../elements/TitleAndBreadcrumb";
-import {currentYear, DateInput} from "../elements/inputs/DateInput";
 import {
     above,
     BOARD_ORDER_NAMES,
@@ -49,27 +42,15 @@ import {
     isAda,
     isDefined,
     isPhy,
-    isEventLeaderOrStaff,
-    Item,
-    itemise,
     matchesAllWordsInAnyOrder,
-    nthHourOf,
     PATHS,
-    selectOnChange,
     siteSpecific,
     useDeviceSize,
     useGameboards,
-    TODAY,
-    UTC_MIDNIGHT_IN_SIX_DAYS,
-    addDays,
-    nthUtcHourOf
 } from "../../services";
-import {Loading} from "../handlers/IsaacSpinner";
-import {AssignmentDTO, GameboardDTO, RegisteredUserDTO, UserGroupDTO} from "../../../IsaacApiTypes";
-import {BoardAssignee, AssignmentBoardOrder, Boards, ActiveModal} from "../../../IsaacAppTypes";
+import {AssignmentDTO, GameboardDTO, RegisteredUserDTO} from "../../../IsaacApiTypes";
+import {BoardAssignee, AssignmentBoardOrder, Boards} from "../../../IsaacAppTypes";
 import {BoardCard} from "../elements/cards/BoardCard";
-import {StyledSelect} from "../elements/inputs/StyledSelect";
-import {PageFragment} from "../elements/PageFragment";
 import {RenderNothing} from "../elements/RenderNothing";
 import {SortItemHeader} from "../elements/SortableItemHeader";
 import {MainContent, SetAssignmentsSidebar, SidebarLayout} from "../elements/layout/SidebarLayout";
@@ -77,244 +58,8 @@ import {HorizontalScroller} from "../elements/inputs/HorizontalScroller";
 import classNames from "classnames";
 import {PromptBanner} from "../elements/cards/PromptBanner";
 import { PageMetadata } from "../elements/PageMetadata";
-
-interface AssignGroupProps {
-    groups: UserGroupDTO[];
-    assignees: BoardAssignee[];
-    board: GameboardDTO | undefined;
-    closeModal: () => void;
-}
-
-const AssignGroup = ({groups, assignees, board, closeModal}: AssignGroupProps) => {
-    const [selectedGroups, setSelectedGroups] = useState<Item<number>[]>([]);
-    const [dueDate, setDueDate] = useState<Date | undefined>(UTC_MIDNIGHT_IN_SIX_DAYS);
-    const [userSelectedDueDate, setUserSelectedDueDate] = useState<boolean>(false);
-    const [scheduledStartDate, setScheduledStartDate] = useState<Date>();
-    const [assignmentNotes, setAssignmentNotes] = useState<string>();
-    const [validationAttempted, setValidationAttempted] = useState<boolean>(false);
-    const user = useAppSelector(selectors.user.loggedInOrNull);
-    const dispatch = useAppDispatch();
-
-    if (!board) return <Loading/>;
-
-    function attemptAssign() {
-        setValidationAttempted(true);
-        if (groupInvalid || dueDateInvalid || startDateInvalid || notesInvalid) {
-            return;
-        }
-        assign();
-    }
-
-    function assign() {
-        dispatch(assignGameboard({
-            boardId: board?.id as string,
-            groups: selectedGroups,
-            dueDate,
-            scheduledStartDate,
-            notes: assignmentNotes,
-            userId: user?.id
-        })).then(success => {
-            if (success) {
-                setSelectedGroups([]);
-                setDueDate(UTC_MIDNIGHT_IN_SIX_DAYS);
-                setUserSelectedDueDate(false);
-                setScheduledStartDate(undefined);
-                setAssignmentNotes('');
-                closeModal();
-            }
-        });
-    }
-
-    function setScheduledStartDateAtSevenAM(e: ChangeEvent<HTMLInputElement>) {
-        const utcDate = e.target.valueAsDate;
-        if (utcDate) {
-            const scheduledDate = new Date(utcDate.getFullYear(), utcDate.getMonth(), utcDate.getDate(), 7);
-            // Sets the scheduled date to 7AM in the timezone of the browser.
-            setScheduledStartDate(scheduledDate);
-
-            if (!userSelectedDueDate) {
-                // Sets the due date to 6 days after the scheduled start date at UTC midnight (if the user hasn't already selected a due date).
-                setDueDate(addDays(6, nthUtcHourOf(0, scheduledDate)));
-            }
-        } else {
-            setScheduledStartDate(null as unknown as Date);
-            {/* DANGER here with force-casting Date|null to Date */}
-        }
-    }
-
-    const yearRange = range(currentYear, currentYear + 5);
-    const dueDateInvalid = isDefined(dueDate) && ((scheduledStartDate ? (nthHourOf(0, scheduledStartDate).valueOf() > dueDate.valueOf()) : false) || TODAY().valueOf() > dueDate.valueOf());
-    const startDateInvalid = scheduledStartDate ? TODAY().valueOf() > scheduledStartDate.valueOf() : false;
-    const groupInvalid = selectedGroups.length === 0 || selectedGroups.some(g => assignees.some(a => a.groupId === g.value));
-    const notesInvalid = isDefined(assignmentNotes) && assignmentNotes.length > 500;
-    
-    //const isAssignmentSetToThisGroup = (group: Item<number>, assignment?: QuizAssignmentDTO) => assignment ? (assignment.quizId === quiz.id && assignment.groupId === group.value && (assignment.dueDate ? assignment.dueDate.valueOf() > Date.now() : true)) : false;
-    //const alreadyAssignedToAGroup = selectedGroups.some(group => quizAssignments?.some(assignment => isAssignmentSetToThisGroup(group, assignment)));
-
-    return <Form onSubmit={(e) => {e.preventDefault(); attemptAssign();}} className="py-2">
-        <FormGroup>
-            <Label data-testid="modal-groups-selector" className="w-100 pb-2">
-                <span className="form-required">Group(s):</span>
-                <div className={classNames({"is-invalid": validationAttempted && groupInvalid})}>
-                    <StyledSelect inputId="groups-to-assign" isMulti isClearable placeholder="None"
-                        value={selectedGroups}
-                        closeMenuOnSelect={false}
-                        onChange={selectOnChange(setSelectedGroups, false)}
-                        options={sortBy(groups, group => group.groupName && group.groupName.toLowerCase()).map(g => itemise(g.id as number, g.groupName))}
-                    />
-                </div>
-                {(selectedGroups.length === 0 
-                    ? <FormFeedback>You must select a group</FormFeedback> 
-                    : <FormFeedback>You cannot reassign an assignment to this group(s) until the due date has passed.</FormFeedback>
-                )}
-            </Label>
-        </FormGroup>
-        <FormGroup>
-            <Label className="w-100 pb-2">
-                <span>Set an optional start date:</span>
-                <DateInput value={scheduledStartDate} placeholder="Select your scheduled start date..."
-                    yearRange={yearRange} invalid={validationAttempted && startDateInvalid}
-                    onChange={setScheduledStartDateAtSevenAM}/>
-                <FormFeedback>{startDateInvalid && "Start date must be in the future."}</FormFeedback>
-            </Label>
-        </FormGroup>
-        <FormGroup>
-            <Label className={"w-100 pb-2"}>
-                <span className="form-required">Due date reminder</span>
-                <DateInput value={dueDate} placeholder="Select your due date..." yearRange={yearRange} invalid={validationAttempted && dueDateInvalid}
-                    onChange={e => { setUserSelectedDueDate(true); setDueDate(e.target.valueAsDate as Date); }}/> {/* DANGER here with force-casting Date|null to Date */}
-                <FormFeedback>{!dueDate && `Since ${siteSpecific("Jan", "January")} 2025, due dates are required for assignments.`}</FormFeedback>
-                <FormFeedback>{dueDateInvalid && "Due date must be on or after start date and in the future."}</FormFeedback>
-            </Label>
-        </FormGroup>
-        <FormGroup>
-            {isEventLeaderOrStaff(user) && <Label className="w-100 pb-2">
-                <span>Notes:</span>
-                <Input type="textarea"
-                    spellCheck={true}
-                    rows={3}
-                    value={assignmentNotes}
-                    onChange={(e: React.ChangeEvent<HTMLInputElement>) => setAssignmentNotes(e.target.value)}
-                />
-                <p className="mt-1 mb-0"><small>{(assignmentNotes || '').length}/500 characters</small></p>
-                <FormFeedback>{notesInvalid && <p className="mt-0 mb-0 text-danger"><small>You have exceeded the maximum length.</small></p> }</FormFeedback>
-            </Label>}
-        </FormGroup>
-        
-        <Button className="my-2" block color={siteSpecific("keyline", "solid")} type="submit">
-            Assign to group{selectedGroups.length > 1 ? "s" : ""}
-        </Button>
-    </Form>;
-};
-
-interface AssignmentDisplayProps {
-    board: GameboardDTO | undefined;
-    assignees: BoardAssignee[];
-    unassignBoard: (props: { boardId: string, groupId: number }) => void;
-}
-
-const AssignmentDisplay = ({board, assignees, unassignBoard}: AssignmentDisplayProps) => {
-    const hasStarted = (a: { startDate?: Date | number }) => !a.startDate || (Date.now() > a.startDate.valueOf());
-
-    const startedAssignees = assignees.filter(hasStarted);
-    const scheduledAssignees = assignees.filter(a => !hasStarted(a));
-
-    function confirmUnassignBoard(groupId: number, groupName?: string) {
-        if (board?.id && confirm(`Are you sure you want to unassign this ${siteSpecific("question deck", "quiz")} from ${groupName ? `group ${groupName}` : "this group"}?`)) {
-            unassignBoard({boardId: board?.id, groupId});
-        }
-    }
-
-    return <> 
-        <div className="py-2 border-bottom d-flex flex-column" data-testid="currently-assigned-to">
-            <span>{siteSpecific("Question deck", "Quiz")} currently assigned to:</span>
-            {startedAssignees.length > 0
-                ? <ul className="p-2 mb-3">{startedAssignees.map(assignee =>
-                    <li data-testid={"current-assignment"} key={assignee.groupId}
-                        className="my-1 px-1 d-flex justify-content-between"
-                    >
-                        <span className="flex-grow-1">{assignee.groupName}</span>
-                        <button 
-                            className="close bg-transparent invert-underline" aria-label="Unassign group" 
-                            onClick={() => confirmUnassignBoard(assignee.groupId, assignee.groupName)}
-                        >
-                            Unassign
-                        </button>
-                    </li>
-                )}</ul>
-                : <p className="px-2">No groups.</p>}
-        </div>
-        <div className="py-2 d-flex flex-column">
-            <span className={classNames("mb-2", siteSpecific("d-flex align-items-center", ""))}>
-                Pending {siteSpecific("assignments", "quiz assignments")}:
-                <i className={siteSpecific("icon icon-info icon-color-grey ms-2", "icon-help mx-1")}
-                    id={`pending-assignments-help-${board?.id}`}/>
-            </span>
-            <UncontrolledTooltip placement="left" autohide={false} target={`pending-assignments-help-${board?.id}`}>
-                These {siteSpecific("assignments", "quizzes")} are scheduled to begin at a future date. On the
-                morning of the scheduled date, students
-                will be able to see the {siteSpecific("assignment", "quiz")}, and will receive a notification email.
-            </UncontrolledTooltip>
-            {scheduledAssignees.length > 0
-                ? <ul className="p-2 mb-3">{scheduledAssignees.map(assignee =>
-                    <li data-testid={"pending-assignment"} key={assignee.groupId}
-                        className="my-1 px-1 d-flex justify-content-between"
-                    >
-                        <span className="flex-grow-1">{assignee.groupName}</span>
-                        {assignee.startDate && <>
-                            <span id={`start-date-${assignee.groupId}`}
-                                className="ms-auto me-2">🕑 {(typeof assignee.startDate === "number"
-                                    ? new Date(assignee.startDate)
-                                    : assignee.startDate).toDateString()}
-                            </span>
-                        </>}
-                        <button className="close bg-transparent" aria-label="Unassign group" onClick={() => confirmUnassignBoard(assignee.groupId, assignee.groupName)}>
-                            ×
-                        </button>
-                    </li>
-                )}</ul>
-                : <p className="px-2">No groups.</p>}
-        </div>
-    </>;
-};
-
-type SetAssignmentsModalProps = {
-    board: GameboardDTO | undefined;
-    assignees: BoardAssignee[];
-    groups: UserGroupDTO[];
-    toggle: () => void;
-    unassignBoard: (props: { boardId: string, groupId: number }) => void;
-};
-
-export const SetAssignmentsModal = (props: SetAssignmentsModalProps): ActiveModal => {
-    const {board, assignees, groups, toggle, unassignBoard} = props;
-
-    return {
-        closeAction: toggle,
-        title: board?.title,
-        body: <SetAssignmentsModalContent
-            board={board} assignees={assignees}
-            groups={groups} toggle={toggle}
-            unassignBoard={unassignBoard}
-        />,
-        buttons: [<Button key={0} color="keyline" className="w-100" onClick={toggle}>Close</Button>]
-    };
-}
-
-const SetAssignmentsModalContent = (props: SetAssignmentsModalProps) => {
-    const {toggle} = props;
-
-    const description = "Scheduled assignments appear to students on the morning of the day chosen, otherwise assignments appear immediately. " +
-        "Assignments are due by the end of the day indicated.";
-
-    return <>
-        <p className="px-1">{description}</p>
-        <hr className="text-center"/>
-        <AssignGroup closeModal={toggle} {...props} />
-        <hr className="text-center"/>
-        <AssignmentDisplay {...props} />
-    </>;
-};
+import { SetAssignmentsModal } from "../elements/modals/SetAssignmentsModal";
+import { PageFragment } from "../elements/PageFragment";
 
 interface SetAssignmentsTableProps {
     user: RegisteredUserDTO;
